fix!: derive unwrapped stream type from input
The passed stream is wrapped in `it-handshake` which can derive the
unrwapped type from the input.  This module assumes TSink will be
`Uint8Array` which may not be the case.

Instead, derive the unwrapped stream type from the input, though it's
meant needing to make the `ProtobufStream` type generic which may
trip up certain use cases (declaring a `ProtobufStream` variable, for
example) so this is a breaking change.

BREAKING CHANGE: ProtobufStream is now generic